Pull KaboodleInner out into a separate module
